Renamed WorkflowResults and WorkflowResultsContext to ProcessingResults (for the general class) and WorkflowResults (for the singleton instance) for consistency, similarly to ProcessingTree and WorkflowTree.
