+++ .github/workflows/unittests.yml
@@ -31,7 +31,11 @@ jobs:
         python -m pip install --upgrade pip
         python -m pip install .
         python -m pip install pytest
-    - name: Test with pytest
+    - name: Run fast tests with pytest
       run: |
         export QT_QPA_PLATFORM=offscreen
         python -m pytest
+    - name: Run slow tests with pytest
+      run: |
+        export QT_QPA_PLATFORM=offscreen
+        python -m pytest -m slow
