+++ src/pydidas/docs/src/manuals/cmdline_apps/cmd_execute_workflow_app.rst
@@ -180,9 +180,9 @@ Accessing results within Python
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 The results from the ExecuteWorkflowApp are written in the global 
-:py:class:`WorkflowResults <pydidas.workflow.WorkflowResultsContext>` (the 
-Singleton instance of :py:class:`WorkflowResults 
-<pydidas.workflow.WorkflowResults>`) which is described in 
+:py:class:`WorkflowResults <pydidas.workflow.WorkflowResults>` (the 
+Singleton instance of :py:class:`ProcessingResults
+<pydidas.workflow.ProcessingResults>`) which is described in
 detail in :ref:`workflow_results`.
 
 List of all ExecuteWorkflowApp Parameters
