+++ src/pydidas/docs/src/manuals/cmdline_global/cmd_workflow_results.rst
@@ -19,7 +19,7 @@ Introduction
 
 The :py:class:`WorkflowResults <pydidas.workflow.WorkflowResults>` is the
 Singleton instance of the
-:py:class:`WorkflowResults <pydidas.workflow.WorkflowResults>`.
+:py:class:`ProcessingResults <pydidas.workflow.ProcessingResults>`.
 It is used for managing the processing results.
 
 This class does not need any configuration as the layout of the results are
@@ -49,7 +49,7 @@ Getting information about the results
 -------------------------------------
 
 To get more information about the results, the
-:py:class:`WorkflowResults <pydidas.workflow.WorkflowResults>`
+:py:class:`WorkflowResults <pydidas.workflow.ProcessingResults>`
 class has several properties to get general information about the stored results
 
 
@@ -104,7 +104,7 @@ Metadata
 
 The metadata of a node ID's Dataset can be accessed using the
 :py:meth:`get_result_metadata(node_id)
-<pydidas.workflow.WorkflowResults.get_result_metadata>`
+<pydidas.workflow.ProcessingResults.get_result_metadata>`
 method. It will return a dictionary with the metadata keys and their respective
 data:
 
@@ -129,12 +129,12 @@ creating a copy of the full data.
 Generic Data
 """"""""""""
 
-The :py:meth:`get_results(node_id) <pydidas.workflow.WorkflowResults.get_results>`
+The :py:meth:`get_results(node_id) <pydidas.workflow.ProcessingResults.get_results>`
 method is available to access the full Dataset with the results of a Plugin.
 The calling parameter is the node ID of the particular Plugin corresponding to
 the results:
 
-.. automethod:: pydidas.workflow.WorkflowResults.get_results
+.. automethod:: pydidas.workflow.ProcessingResults.get_results
     :noindex:
 
 An example is given below:
@@ -191,11 +191,11 @@ Flattened scan dimensions
 For some applications, it might be interesting to ignore the detailed shape of
 the scan and flatten the scan to a *timeline*. The
 :py:meth:`get_results_for_flattened_scan(node_id)
-<pydidas.workflow.WorkflowResults.get_results_for_flattened_scan>`
+<pydidas.workflow.ProcessingResults.get_results_for_flattened_scan>`
 method allows to get a Dataset with all the scan dimensions flattened to a
 single dimension renamed to *timeline*:
 
-.. automethod:: pydidas.workflow.WorkflowResults.get_results_for_flattened_scan
+.. automethod:: pydidas.workflow.ProcessingResults.get_results_for_flattened_scan
     :noindex:
 
 An example is given below:
@@ -249,13 +249,13 @@ Accessing a data subset
 For convenience, a method to access only a subset of the data is implemented as
 well:
 
-.. automethod:: pydidas.workflow.WorkflowResults.get_result_subset
+.. automethod:: pydidas.workflow.ProcessingResults.get_result_subset
     :noindex:
 
 This method is interesing if the user wants to access a specific subset in the
 flattened data, for example the results for the frames 40 to 55 of the
 experiment. This can easily be done using the :py:meth:`get_result_subset
-<pydidas.workflow.WorkflowResults.get_result_subset>`
+<pydidas.workflow.ProcessingResults.get_result_subset>`
 method, as demonstrated in the example below:
 
 
@@ -305,11 +305,11 @@ Saving results
 --------------
 
 Saving the results is achieved via the :py:meth:`save_results_to_disk
-<pydidas.workflow.WorkflowResults.save_results_to_disk>`
+<pydidas.workflow.ProcessingResults.save_results_to_disk>`
 method:
 
 .. automethod::
-    pydidas.workflow.WorkflowResults.save_results_to_disk
+    pydidas.workflow.ProcessingResults.save_results_to_disk
     :noindex:
 
 For now, the only available saver is 'HDF5' and additional savers will be added
