+++ tests/workflow/test_workflow_result_selector.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2025, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -18,7 +18,7 @@
 """Unit tests for pydidas modules."""
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2025, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -35,16 +35,16 @@ from pydidas.contexts import ScanContext
 from pydidas.core import Dataset, Parameter, UserConfigError
 from pydidas.plugins import PluginCollection
 from pydidas.unittest_objects import DummyLoader, DummyProc, DummyProcNewDataset
-from pydidas.workflow import WorkflowResultsContext, WorkflowTree
-from pydidas.workflow.result_io import WorkflowResultIoMeta
+from pydidas.workflow import WorkflowResults, WorkflowTree
+from pydidas.workflow.result_io import ProcessingResultIoMeta
 from pydidas.workflow.workflow_results_selector import WorkflowResultsSelector
 
 
 PLUGINS = PluginCollection()
 SCAN = ScanContext()
 TREE = WorkflowTree()
-RES = WorkflowResultsContext()
-SAVER = WorkflowResultIoMeta
+RES = WorkflowResults()
+SAVER = ProcessingResultIoMeta
 
 
 class TestWorkflowResultSelector(unittest.TestCase):
